archive/requirements.txt
-e .
numpy
scipy
opencv-python
torchdiffeq
